refactored, added variable metadata and dep/indep var names to FAIR dict
