+++ acro/acro_tables.py
@@ -20,20 +20,30 @@ from pandas.api.types import CategoricalDtype
 import copy
 
 from . import utils
-from .checks import SDCChecks
+from .checks import SDCChecks, ChecksResults, ManyChecksResults
 from .record import Records
 from .aggregationfunctions import agg_mode, agg_values_are_same
 from .aggregationfunctions import agg_negative,agg_missing
 from .aggregationfunctions import agg_p_percent, agg_nk, agg_threshold
 
+from .constants import SDMX_DIMENSION, SDMX_MEASURE
 
 
+from .table_utils import (
+    aggfunc_to_strings,
+    axis_to_list, 
+    collate_risk_assessments,
+    get_axis_metadata,
+    get_redacted_table,
+    get_debugging_table_analysis,
+    get_variable_metadata,
+    get_variable_type_dict
+
+)
 
 logger = logging.getLogger("acro")
 
-#https://www.sdmx.io/learning/essential-sdmx-structural-modelling/003-concept-mapping/
-SDMX_DIMENSION:str= "sdmx.model.Dimension"
-SDMX_MEASURE:str= "sdmx.model.Measure"
+
 
 
 AGGFUNC: dict[str, str | Callable] = {
@@ -45,16 +55,7 @@ AGGFUNC: dict[str, str | Callable] = {
     "mode": agg_mode,
 }
 
-AGGFUNC_TO_TYPE: dict[str, str] = {
-    "count": "FrequencyTable",
-    "mode": "Mode",
-    "median": "Median",
-    "mean": "Mean",
-    "std": "StandardDeviation",
-    "sum": "Sum",
-    "min": "Minimum",
-    "max": "Maximum",
-}
+
 
 
 # aggregation function parameters
@@ -132,6 +133,7 @@ class Tables:
             - If passed 'columns' will normalize over each column.
             - If margins is `True`, will also normalize margin values.
         show_suppressed : bool. default False
+            Deprecated in v.10, only present for backwards compatability
             how the totals are being calculated when the suppression is true
 
         Returns
@@ -141,6 +143,9 @@ class Tables:
         """
         logger.debug("crosstab()")
         command: str = utils.get_command("crosstab()", stack())
+        _ = show_suppressed #hide complaint about unused legacy variable
+
+        
         # syntax checking
         if aggfunc is not None:
             if values is None or isinstance(values, list):
@@ -154,25 +159,8 @@ class Tables:
         index = axis_to_list(index)
         columns= axis_to_list(columns)
         
-        #create data dictionary
-        #TODO handle arraylike as well as series
-        #TODO handle rownmanes/colnames
-        variable_metadata:dict[str,dict] = {}
-        variable_metadata.update(get_catvar_data(index,where="rows"))
-        variable_metadata.update(get_catvar_data(columns,where="columns"))
-        if values is not None:
-            name= values.name if isinstance(values,Series) else "unknown_measure"
-            variable_metadata[name]= {"location": "cells",
-                                  "sequence_id": 0,
-                                  "dtype": values[0].dtype,
-                                  "type" : SDMX_MEASURE,
-                                  "categories":[]
-                                 }
-  
-
-
         
-        # save list and dict to reduce clutter
+        # save list and dict to reduce code clutter
         args = (index, columns)
         kwargs = {
             "values": values,
@@ -186,118 +174,39 @@ class Tables:
         }
         # enforce dropna=False for SDC reasons
         kwargs["dropna"] = False
-
-        _ = show_suppressed
-        # suppression_args = {"show_suppressed": show_suppressed}
+        variable_metadata = get_variable_metadata(index, columns,values)
+        #make object containing these that can easily be passed around
+        model: dict[str, Any] = {
+            "args": args, 
+            "kwargs": kwargs,
+            "variable_metadata":variable_metadata,
+            "risk_appetite":self.sdc_checks.risk_appetite
+                                }
 
         ## requested table
         table: DataFrame = pd.crosstab(*args, **kwargs)
-
-        ## get names of Aggregation Functions- sole case first
-        analysis_names: list[str] = []
-        table_width_multiplier=1
-
-        if aggfunc is None:
-            analysis_names.append(AGGFUNC_TO_TYPE.get("count", "missing"))
-        if isinstance(aggfunc, str):
-            analysis_names.append(AGGFUNC_TO_TYPE.get(aggfunc, "missing"))
-        if isinstance(aggfunc, list):
-            for i in aggfunc:
-                analysis_names.append(AGGFUNC_TO_TYPE.get(i, "missing"))
-                table_width_multiplier+=1
+        analysis_names: list[str] = aggfunc_to_strings(aggfunc)
 
         ## run the checks and get the masks
-        allmasks: dict[str, DataFrame] = {}
-        fair_dicts: dict[str, dict] = {}
-        statuses: dict[str, str] = {}
-        summaries: dict[str, str] = {}
-
-        model: dict[str, Any] = {"args": args, "kwargs": kwargs}
-
+        collatedres=ManyChecksResults()
         for analysis in analysis_names:
-            status, summary, masks, fair_dict = self.sdc_checks.run_checks_for_analysis(
-                analysis_names[0], model
+            collatedres.allchecksresults[analysis] = self.sdc_checks.run_checks_for_analysis(
+                analysis, model
             )
 
-            summaries[analysis] = summary
-            statuses[analysis] = status
-            allmasks[analysis] = masks
-            #TODO should check if there are duplicate keys with different values
-            fair_dicts[analysis] = fair_dict
+        logging.debug(get_debugging_table_analysis(collatedres.allchecksresults))
 
-        collated_assessment = collate_risk_assessments(table, allmasks)
+        collated_assessment = collate_risk_assessments(table, collatedres.allchecksresults)
+        sdc_details: dict = collatedres.get_table_sdc()
+        overall_status:str = collatedres.get_overall_status()
+        allsummary:str=collatedres.get_overall_summary()
+        fair_dict = collatedres.get_overall_fair()
+        fair_dict.update( get_variable_type_dict(variable_metadata ))
 
-        ## Debugging print out
-        debug_print = False
-        if debug_print:
-            print("\n====start acro.crosstab print statement=====")
 
-            print("\n== statuses==\n")
-            for key, val in statuses.items():
-                print(f" {key} : {val}")
 
-            print("\n== summaries==\n")
-            for key, val in summaries.items():
-                print(f" {key} : {val}")
-
-            print("\n== allmasks==\n")
-            for name, mask in allmasks.items():
-                print(f"\nMask for {name}")
-                for key, val in mask.items():
-                    print(f"{key} \n{val}")
-
-            print("\n== fair_dicts==\n")
-            for key, val in fair_dicts.items():
-                if isinstance(val, dict):
-                    for key2, val2 in val.items():
-                        print(f" {key2} : {val2}")
-                else:
-                    print(f" {key} : {val}")
-
-            print("\n=== collated masks ===")
-            print(f"{collated_assessment}")
-            print("====end acro.crosstab print statement=====\n")
-
-        ## summarise
-        sdc_details: dict = get_table_sdc(allmasks)
-        if "fail" in statuses:
-            overall_status = "fail"
-        elif "review" in statuses:
-            overall_status = "review"
-        else:
-            overall_status = "pass"
-
-        allsummary=""
-        for key, val in summaries.items():
-            allsummary += f'\n{key} : {val}'
-
-        #summary = "".join(summaries)
-        #logger.info(f'summary just before applying suppression is a {type(summary)}'
-        #           f' with contents {allsummary}'
-        #          )
-        # # apply the suppression
         if self.suppress:
-            queries: list[str] = get_queries_from_collated_risk(collated_assessment,aggfunc)
-            relevant_data:DataFrame = get_relevant_dataframe(args,kwargs,variable_metadata)
-            redacted_data: DataFrame = get_redacted_data(relevant_data,queries)
-            #ensure missing categories are present
-            for name in list(redacted_data):
-                if variable_metadata[name]["type"]==SDMX_DIMENSION:
-                    cat_type= CategoricalDtype(
-                            categories=variable_metadata[name]['categories'],
-                            ordered=variable_metadata[name]['ordered']
-                        )
-                    redacted_data[name]=redacted_data[name].astype(cat_type)
-
-
-            newargs = translate_args_to_newdf(args,redacted_data)
-            newkwargs:dict[str:Any]= copy.deepcopy(kwargs)
-            newkwargs['dropna']=False
-            if "values" in kwargs.keys() and kwargs["values"] is not None:
-                newkwargs["values"]= redacted_data[ kwargs["values"].name]
-            table = pd.crosstab(*newargs,**newkwargs)
-            if self.sdc_checks.risk_appetite["zeros_are_disclosive"]:
-                table.replace({0:np.nan},inplace=True)
+            table=get_redacted_table(model, collated_assessment)
 
         # record output
         self.results.add(
@@ -305,7 +214,7 @@ class Tables:
             output_type="table",
             properties={"method": "crosstab"},
             sdc=sdc_details,
-            fair=fair_dicts,
+            fair=fair_dict,
             command=command,
             summary=allsummary,
             outcome=collated_assessment,
@@ -1085,1395 +994,7 @@ def _rounded_survival_table(
 
 
 
-def axis_to_list(axis:Series|list[Series])->list[Series]:
-    """translate axis into standard format
 
 
-    Convert variables describing an axis (row/column) into a list
-    to simplify code. Wraps the  input inside a list if it   is a single series
-    or leaves it unchanged if it is already a list of series
-    
-    
-    Parameters:
-    -----------
-    axis: pandas series or list of series
 
-    Returns
-    -------
-    list [Series]
- 
-    """
-    if not isinstance(axis,list):
-        foo:list= []
-        foo.append(axis)
-        return foo
-    return axis
 
-
-
-
-
-
-def collate_risk_assessments(
-    table: DataFrame, 
-    allmasks: dict[str, DataFrame]
-    ) -> DataFrame:
-    """Collate the Risk Assessment for a table.
-
-    Parameters
-    ----------
-    table : DataFrame
-        Table to be risk assessed.
-    masks : dict[str, DataFrame]
-        Dictionary of tables specifying individual risk assessment masks .
-
-    Returns
-    -------
-    DataFrame
-        Table with outcomes of suppression checks.
-    """
-
-    outcome_df = DataFrame(index=table.index, columns=table.columns)
-    #drop column repetitions for multiple aggregatino functions
-    lowestlevelfound = []
-    to_drops=[]
-    for thetuple in list(outcome_df):
-        if thetuple[-1] in lowestlevelfound:
-            to_drops.append(thetuple)
-        else:
-            lowestlevelfound.append(thetuple[-1])
-    for drop in to_drops:
-        outcome_df= outcome_df.drop(drop,axis='columns')
-
-
-    outcome_df.fillna("", inplace=True)
-
-    checks_seen:list[str] = []
-    for analysis, masks in allmasks.items():
-        # report if negatives are present
-        if "negative" in masks:
-            mask = masks["negative"]
-            outcome_df[mask.values] = "negative"
-        #   report if missing values are present
-        elif "missing" in masks:
-            mask = masks["missing"]
-            outcome_df[mask.values] = "missing"
-        # collate at-risk cells from individual risk masks
-        else:
-            for name, mask in masks.items():
-                if name in checks_seen:
-                    continue
-                checks_seen.append(name)
-                try:
-                    tmp_df = DataFrame(index=outcome_df.index, columns=outcome_df.columns)
-                    tmp_df.fillna("", inplace=True)
-                    tmp_df[mask.values] = name + "; "
-                    outcome_df += tmp_df
-                except TypeError:
-                    logger.warning("problem mask %s is not binary", name)
-                except ValueError as error:  # pragma: no cover
-                    error_message = (
-                        f"An error occurred with the following details"
-                        f":\n Name: {name}\n "
-                        f"Mask of size {mask.shape}"
-                        f"table of shape {table.shape}"
-                    )
-                    raise ValueError(error_message) from error
-    
-        outcome_df = outcome_df.replace({"": "ok"})
-    logger.info("outcome_df:\n%s", utils.prettify_table_string(outcome_df))
-    return outcome_df
-
-def get_catdtype(series:pd.Series)->pd.CategoricalDtype:
-    """convert axis dimension to CategoricalDtype"""
-    ordered=  True if series.astype(int, errors="ignore").dtype=='int64' else False
-    categories= np.sort(series.unique())
-    cat_type= pd.CategoricalDtype(categories, ordered)
-    return cat_type
-
-def get_catvar_data(axis,where:str)-> dict:
-    """ get  metadata for categorical variables describing an axis
-    
-    Cycle through the categorical variables that define an axis 
-    and construct a meta data dictionary describing them 
-    
-    Parameters
-    ----------
-    axis: list[pd.Series]
-        list of series defining a dimension in an analysis
-    where: str
-        axis reference i.e. "rows" or "columns"
-
-    Returns
-    -------
-
-    dict   
-        one entry for item in list provided
-        key is name of series
-        dict of values describe location, type, categories present
-        
-    """
-    
-    metadata: dict[str,dict] = {}
-    for idx,dimension in enumerate(axis):
-        entry:dict={}
-        if not isinstance(dimension,Series):
-            logger.info("unable to construct meta data for "
-                         " component of %s that is not a pandas series", where)
-        else:
-            name = dimension.name
-            cat_type=get_catdtype(dimension)
-            metadata[name]= {
-                "location": where,
-                "sequence_id": idx,
-                "dtype": str(cat_type.categories.dtype),
-                "type" : SDMX_DIMENSION,
-                "categories":list(cat_type.categories),
-                "ordered": cat_type.ordered 
-            }
-    return metadata
-
-    
-def get_table_sdc(allmasks: dict[str, dict]) -> dict[str, Any]:
-    """Return the SDC dictionary using the suppression masks.
-
-    Parameters
-    ----------
-    allmasks : dict[str, dict]
-        Nested Dictionary of tables specifying suppression masks for application.
-    """
-    # summary of number of cells to be suppressed
-    sdc: dict[str, Any] = {"summary": {}, "cells": {}}
-
-    checks_seen:list[str] = []
-    for analysis, masks in allmasks.items():
-        for name, mask in masks.items():
-            if name in checks_seen:
-                continue
-            checks_seen.append(name)
-            sdc["summary"][name] = int(np.nansum(mask.to_numpy()))
-            sdc["cells"][name] = []
-            # positions of cells to be suppressed
-            true_positions = np.column_stack(np.where(mask.values))
-            for pos in true_positions:
-                row_index, col_index = pos
-                sdc["cells"][name].append([int(row_index), int(col_index)])
-    return sdc
-
-
-def get_summary(sdc: dict[str, Any]) -> tuple[str, str]:
-    """Return the status and summary of the suppression masks.
-
-    Parameters
-    ----------
-    sdc : dict
-        Properties of the SDC checks.
-
-    Returns
-    -------
-    str
-        Status: {"review", "fail", "pass"}.
-    str
-        Summary of the suppression masks.
-    """
-    status: str = "pass"
-    summary: str = ""
-    sdc_summary = sdc["summary"]
-    sup: str = "suppressed" if sdc_summary["suppressed"] else "may need suppressing"
-    if sdc_summary["negative"] > 0:
-        summary += "negative values found"
-        status = "review"
-    elif sdc_summary["missing"] > 0:
-        summary += "missing values found"
-        status = "review"
-    else:
-        if sdc_summary["threshold"] > 0:
-            summary += f"threshold: {sdc_summary['threshold']} cells {sup}; "
-            status = "review" if sdc_summary["suppressed"] else "fail"
-        if sdc_summary["p-ratio"] > 0:
-            summary += f"p-ratio: {sdc_summary['p-ratio']} cells {sup}; "
-            status = "review" if sdc_summary["suppressed"] else "fail"
-        if sdc_summary["nk-rule"] > 0:
-            summary += f"nk-rule: {sdc_summary['nk-rule']} cells {sup}; "
-            status = "review" if sdc_summary["suppressed"] else "fail"
-        if sdc_summary["all-values-are-same"] > 0:
-            summary += (
-                f"all-values-are-same: {sdc_summary['all-values-are-same']} "
-                f"cells {sup}; "
-            )
-            status = "review" if sdc_summary["suppressed"] else "fail"
-    if summary != "":
-        summary = f"{status}; {summary}"
-    else:
-        summary = status
-    logger.info("get_summary(): %s", summary)
-    return status, summary
-
-
-def add_backticks(name: str) -> str:
-    """Add backticks to a name if it contains spaces and doesn't have them.
-
-    Parameters
-    ----------
-    name : str
-        The name to add backticks to.
-
-    Returns
-    -------
-    str
-        The name with backticks if needed.
-    """
-    if isinstance(name, str) and " " in name and not name.startswith("`"):
-        return f"`{name}`"
-    return name  # pragma: no cover
-
-
-def _format_label_condition(level_names: list[Any], label: Any) -> list[str]:
-    """Format a label into a list of condition strings.
-
-    Parameters
-    ----------
-    level_names : list
-        The names of the levels.
-    label : tuple or scalar
-        The label value(s).
-
-    Returns
-    -------
-    list[str]
-        List of condition strings for this label.
-    """
-    parts = []
-    if isinstance(label, tuple):
-        for level, val in zip(level_names, label, strict=False):
-            level = add_backticks(str(level))
-            if isinstance(val, (int, float)):
-                parts.append(f"({level} == {val})")
-            else:
-                parts.append(f'({level} == "{val}")')
-    else:
-        level = add_backticks(str(level_names[0]))
-        if isinstance(label, (int, float)):
-            parts.append(f"({level} == {label})")
-        else:
-            parts.append(f'({level} == "{label}")')
-    return parts
-
-
-def get_relevant_dataframe(args:list,kwargs:dict[str:Any], metadata:dict)->DataFrame:
-    """ extract copy of data relevant to crosstab into new DataFrame
-
-    Assumes preprocessing has happenededindex and columns in args should both have been converted into lists of Series
-    
-    
-    Parameters
-    ----------
-    args:list[str|list]
-        list of index, columns from call to crosstab function
-        should have already been converted to lists
-    kwargs: dict
-        kwargs for crosstab function
-    metadata: dict
-        information for creating categorical datatypes so as to preserve possible values
-
-    Returns
-    -------
-    dataframe containing copies of pandas series need to calculate the  crosstab
-    """
-    series_list:list= []
-    if "values" in kwargs.keys() and kwargs["values"] is not None:
-        series_list.append( kwargs["values"].copy())
-    if not (isinstance (args,tuple) and len(args)==2):
-        print(f'args is of type {type(args)} and contents {args}\n')
-        raise ValueError("list passed as positional args has wrong type or length")
-    for contents in args:
-        if not isinstance(contents,list):
-            raise TypeError("index and columns should be lists")
-        for series in contents:
-            series_list.append(series.copy())
-
-    relevant_data = DataFrame(series_list).T
-    return relevant_data
-
-
-
-def translate_args_to_newdf(arguments:list,redacted_data:DataFrame)->list:
-    """translate arguments or keys from one data frame to another
-
-    Parameters:
-    -----------
-    arguments: list
-        list of positional arguments to be translated to a different dataframe
-    redacted_data: Dataframe
-        the name of the 'host' dataframe
-        
-
-    Returns:
-    ---------
-    list
-         arguments  translate on to columns with the same name in the host DataFrame 
-    """
-    #todo put in checks to make this robust
-    #decide whether to return args i.e. don't do redaction/suppression
-    #instead of raising valueerror
-    newargs:list = []
-    if not (isinstance (arguments,tuple) and len(arguments)==2):
-        raise ValueError("list passed as positional args has wrong type or length")
-    for contents in arguments:
-        if isinstance (contents,pd.Series):
-            newargs.append(redacted_data[contents.name])
-        elif isinstance(contents,list):
-            newlist:list=[]
-            for series in contents:
-                newlist.append(redacted_data[series.name])
-            newargs.append(newlist)
-    return newargs
-
-        
-    
-def _get_cell_query(
-    mask: DataFrame,
-    row_index: int,
-    col_index: int,
-    index_level_names: list[Any],
-    column_level_names: list[Any],
-) -> str | None:
-    """Generate a query string for a cell if it's marked as true in the mask.
-
-    Parameters
-    ----------
-    mask : DataFrame
-        The suppression mask.
-    row_index : int
-        Row index.
-    col_index : int
-        Column index.
-    index_level_names : list
-        Names of index levels.
-    column_level_names : list
-        Names of column levels.
-
-    Returns
-    -------
-    str or None
-        Query string if cell is true, None otherwise.
-    """
-    if not mask.iloc[row_index, col_index]:
-        return None
-
-    parts = []
-    row_label = mask.index[row_index]
-    col_label = mask.columns[col_index]
-
-    parts.extend(_format_label_condition(index_level_names, row_label))
-    parts.extend(_format_label_condition(column_level_names, col_label))
-
-    return " & ".join(parts)
-
-
-def get_queries_from_collated_risk(
-    collated_risk: DataFrame,
-    aggfunc:str|None
-) -> list[str]:
-    """Return a list of the boolean conditions for each true (disclosive) cell in the suppression masks.
-
-    Parameters
-    ----------
-    masks : dict[str, DataFrame]
-        Dictionary of tables specifying suppression masks for application.
-    aggfunc : str | None
-        The aggregation function
-
-    Returns
-    -------
-    str
-        The boolean conditions for each true (disclosive) cell in the suppression masks.
-    """
-    true_cell_queries = []
-    themask = collated_risk.copy()
-
-    themask=themask.replace({"ok":False})
-    themask = themask.mask(themask != False, other=True)
-    if aggfunc is not None:
-        if themask.columns.nlevels > 1:
-            themask = themask.droplevel(0, axis=1)
-    index_level_names = themask.index.names
-    column_level_names = themask.columns.names
-    for col_index, _ in enumerate(themask.columns):
-        for row_index, _ in enumerate(themask.index):
-            query = _get_cell_query(
-                themask, row_index, col_index, index_level_names, column_level_names
-            )
-            if query is not None:
-                true_cell_queries.append(query)
-    true_cell_queries = list(set(true_cell_queries))
-    return true_cell_queries
-
-
-def get_redacted_data(data:DataFrame,queries:list[str])->DataFrame:
-    """Apply set of queries to remove sensitive data from  DataFrame
-
-    Parameters:
-    -------------
-    data: pandas DataFrame
-        the raw data
-    queries: list[str]
-        a set of queries that define the data in cells marked as being disclosive
-
-    Returns:
-    ---------
-    DataFrame 
-         the data after the sensitive data has been removed
-    """
-    redacted_data = data.copy()
-    #logger.info(f'data has shape {data.shape}')
-    for query in queries:
-        #logger.info(f'applying query{query}')
-        redacted_data.query(f"not ({query})",inplace=True)
-        #logger.info(f'now redacted data has shape {redacted_data.shape}')
-
-    if not set(data.columns)==set(redacted_data.columns):
-        logger.warning("Error created redacted data"
-                       " - unable to apply suppression"
-                      )
-        return data
-    return redacted_data
-
-
-
-
-def unused_get_index_columns(
-    index: Any, columns: Any, data: DataFrame
-) -> tuple[list[Any] | Series, list[Any] | Series]:
-    """Get the index and columns from the data dataframe.
-
-    Parameters
-    ----------
-    index : array-like, Series, or list of arrays/Series
-        Values to group by in the rows.
-    columns : array-like, Series, or list of arrays/Series
-        Values to group by in the columns.
-    data : dataframe
-        Table of the index and columns combined.
-
-    Returns
-    -------
-    List | Series
-        The index extracted from the data.
-    List | Series
-        The columns extracted from the data.
-    """
-    shift = 1
-    if isinstance(index, list):
-        index_new = []
-        for i in range(len(index)):
-            index_new.append(data.iloc[:, i])
-        shift = len(index)
-    else:
-        index_new = data[index.name]
-
-    if isinstance(columns, list):
-        columns_new = []
-        for i in range(shift, shift + len(columns)):
-            columns_new.append(data.iloc[:, i])
-    else:
-        columns_new = data[columns.name]
-    return index_new, columns_new
-
-
-def unused_crosstab_with_totals(  # pylint: disable=too-many-arguments,too-many-locals
-    masks: dict[str, DataFrame],
-    aggfunc: Any,
-    index: Any,
-    columns: Any,
-    values: Any,
-    margins: bool,
-    margins_name: str,
-    dropna: bool,
-    crosstab: bool,
-    rownames: Any = None,
-    colnames: Any = None,
-    normalize: bool | str = False,
-    data: DataFrame | None = None,
-    fill_value: Any = None,
-    observed: bool = False,
-    sort: bool = False,
-) -> DataFrame | None:
-    """Recalculate the crosstab table when margins are true and suppression is true.
-
-    Parameters
-    ----------
-    masks : dict[str, DataFrame]
-        Dictionary of tables specifying suppression masks for application.
-    aggfunc : str | None
-        The aggregation function.
-    index : array-like, Series, or list of arrays/Series
-        Values to group by in the rows.
-    columns : array-like, Series, or list of arrays/Series
-        Values to group by in the columns.
-    index : array-like, Series, or list of arrays/Series
-            Values to group by in the rows.
-    columns : array-like, Series, or list of arrays/Series
-        Values to group by in the columns.
-    values : array-like, optional
-        Array of values to aggregate according to the factors.
-        Requires `aggfunc` be specified.
-    rownames : sequence, default None
-        If passed, must match number of row arrays passed.
-    colnames : sequence, default None
-        If passed, must match number of column arrays passed.
-    aggfunc : str, optional
-        If specified, requires `values` be specified as well.
-    margins : bool, default False
-        Add row/column margins (subtotals).
-    margins_name : str, default 'All'
-        Name of the row/column that will contain the totals
-        when margins is True.
-    dropna : bool, default True
-        Do not include columns whose entries are all NaN.
-    normalize : bool, {'all', 'index', 'columns'}, or {0,1}, default False
-        Normalize by dividing all values by the sum of values.
-        - If passed 'all' or `True`, will normalize over all values.
-        - If passed 'index' will normalize over each row.
-        - If passed 'columns' will normalize over each column.
-        - If margins is `True`, will also normalize margin values.
-
-    Returns
-    -------
-    DataFrame
-        Crosstabulation of data
-    """
-    true_cell_queries = get_queries(masks, aggfunc)
-    if crosstab:
-        data = create_dataframe(index, columns)
-    if data is None:
-        raise AssertionError("data must be set when applying crosstab queries")
-    for query in true_cell_queries:
-        data = data.query(f"not ({query})")
-
-    # get the index and columns from the data after the queries are applied
-    try:
-        if crosstab:
-            index_new, columns_new = get_index_columns(index, columns, data)
-            # apply the crosstab with the new index and columns
-            table = pd.crosstab(
-                index_new,
-                columns_new,
-                values=values,
-                rownames=rownames,
-                colnames=colnames,
-                aggfunc=aggfunc,
-                margins=margins,
-                margins_name=margins_name,
-                dropna=dropna,
-                normalize=normalize,
-            )
-
-            if table.empty:
-                raise ValueError("empty table")
-
-            table, _ = delete_empty_rows_columns(table)
-            masks = create_crosstab_masks(
-                index_new,
-                columns_new,
-                values,
-                rownames,
-                colnames,
-                aggfunc,
-                margins,
-                margins_name,
-                dropna,
-                normalize,
-            )
-
-            # Force the apply_suppression not to display the outcome dataframe
-            previous_level = logger.getEffectiveLevel()
-            logger.setLevel(logging.WARNING)
-            # apply the suppression
-            table, _ = apply_suppression(table, masks)
-            logger.setLevel(previous_level)
-
-        else:
-            table = pd.pivot_table(
-                data=data,
-                values=values,
-                index=index,
-                columns=columns,
-                aggfunc=aggfunc,
-                fill_value=fill_value,
-                margins=margins,
-                dropna=dropna,
-                margins_name=margins_name,
-                observed=observed,
-                sort=sort,
-            )
-
-    except ValueError:
-        logger.warning(
-            "All the cells in this data are disclosive."
-            " Thus suppression can not be applied"
-        )
-        return None
-    return table
-
-
-def unused_manual_crossstab_with_totals(  # pylint: disable=too-many-arguments
-    table: DataFrame,
-    aggfunc: str | list[str] | None,
-    index: Any,
-    columns: Any,
-    values: Any,
-    rownames: Any,
-    colnames: Any,
-    margins: bool,
-    margins_name: str,
-    dropna: bool,
-    normalize: bool | str,
-) -> DataFrame | None:
-    """Recalculate the crosstab table when margins are true and suppression is true.
-
-    Parameters
-    ----------
-    table : Dataframe
-        The suppressed table.
-    aggfunc : str | None
-        The aggregation function.
-    index : array-like, Series, or list of arrays/Series
-            Values to group by in the rows.
-    columns : array-like, Series, or list of arrays/Series
-        Values to group by in the columns.
-    values : array-like, optional
-        Array of values to aggregate according to the factors.
-        Requires `aggfunc` be specified.
-    rownames : sequence, default None
-        If passed, must match number of row arrays passed.
-    colnames : sequence, default None
-        If passed, must match number of column arrays passed.
-    margins : bool, default False
-        Add row/column margins (subtotals).
-    margins_name : str, default 'All'
-        Name of the row/column that will contain the totals
-        when margins is True.
-    dropna : bool, default True
-        Do not include columns whose entries are all NaN.
-    normalize : bool, {'all', 'index', 'columns'}, or {0,1}, default False
-        Normalize by dividing all values by the sum of values.
-        - If passed 'all' or `True`, will normalize over all values.
-        - If passed 'index' will normalize over each row.
-        - If passed 'columns' will normalize over each column.
-        - If margins is `True`, will also normalize margin values.
-
-    Returns
-    -------
-    DataFrame
-        Crosstabulation of data
-    """
-    if isinstance(aggfunc, list):
-        logger.warning(
-            "We can not calculate the margins with a list of aggregation functions. "
-            "Please create a table for each aggregation function"
-        )
-        return None
-    if aggfunc is None or aggfunc == "sum" or aggfunc == "count":
-        table = recalculate_count_margins(table, margins_name)
-
-    elif aggfunc == "mean":
-        count_table = pd.crosstab(
-            index=index,
-            columns=columns,
-            values=values,
-            rownames=rownames,
-            colnames=colnames,
-            aggfunc="count",
-            margins=margins,
-            margins_name=margins_name,
-            dropna=dropna,
-            normalize=normalize,
-        )
-        # suppress the cells in the count by mimicking the suppressed cells in the table
-        count_table = count_table.where(table.notna(), other=np.nan)
-        # delete any columns from the count_table that are not in the table
-        columns_to_keep = table.columns
-        count_table = count_table[columns_to_keep]
-        if count_table.index.is_numeric():  # pragma: no cover
-            count_table = count_table.sort_index(axis=1)
-        # recalculate the margins considering the nan values
-        count_table = recalculate_count_margins(count_table, margins_name)
-        # multiply the table by the count table
-        table[margins_name] = 1
-        table.loc[margins_name, :] = 1
-        multip_table = count_table * table
-        if multip_table.index.is_numeric():  # pragma: no cover
-            multip_table = multip_table.sort_index(axis=1)
-        # calculate the margins columns
-        table[margins_name] = (
-            multip_table.drop(margins_name, axis=1).sum(axis=1)
-            / multip_table[margins_name]
-        )
-        # calculate the margins row
-        if not isinstance(count_table.index, pd.MultiIndex):  # single row
-            table.loc[margins_name, :] = (
-                multip_table.drop(margins_name, axis=0).sum()
-                / multip_table.loc[margins_name, :]
-            )
-        else:  # multiple rows
-            table.loc[(margins_name, ""), :] = (
-                multip_table.drop(margins_name, axis=0).sum()
-                / multip_table.loc[(margins_name, ""), :]
-            )
-        # calculate the grand margin
-        if not isinstance(count_table.columns, pd.MultiIndex) and not isinstance(
-            count_table.index, pd.MultiIndex
-        ):  # single column, single row
-            table.loc[margins_name, margins_name] = (
-                multip_table.drop(index=margins_name, columns=margins_name).sum().sum()
-            ) / multip_table.loc[margins_name, margins_name]
-        else:  # multiple columns or multiple rows
-            table.loc[margins_name, margins_name] = (
-                multip_table.drop(index=margins_name, columns=margins_name).sum().sum()
-            ) / multip_table.loc[margins_name, margins_name][0]
-
-    elif aggfunc == "std":
-        table = table.drop(margins_name, axis=1)
-        table = table.drop(margins_name, axis=0)
-        logger.warning(
-            "The margins with the std agg func can not be calculated. "
-            "Please set the show_suppressed to false to calculate it."
-        )
-        return table
-    return table
-
-
-def unused_recalculate_count_margins(table: DataFrame, margins_name: str) -> DataFrame:
-    """Recalculate the margins in a table of counts.
-
-    Parameters
-    ----------
-    table : Dataframe
-        The suppressed table.
-    margins_name : str, default 'All'
-        Name of the row/column that will contain the totals
-
-    Returns
-    -------
-    DataFrame
-        Table with new calculated margins
-    """
-    table = table.drop(margins_name, axis=1)
-    rows_total = table.sum(axis=1)
-    table.loc[:, margins_name] = rows_total
-    if isinstance(table.index, pd.MultiIndex):
-        table = table.drop(margins_name, axis=0)
-        cols_total = table.sum(axis=0)
-        table.loc[(margins_name, ""), :] = cols_total
-    else:
-        table = table.drop(margins_name, axis=0)
-        cols_total = table.sum(axis=0)
-        table.loc[margins_name] = cols_total
-    return table
-
-
-def unused_create_crosstab_masks(  # pylint: disable=too-many-arguments,too-many-locals
-    index: Any,
-    columns: Any,
-    values: Any,
-    rownames: Any,
-    colnames: Any,
-    agg_func: str | Callable | list[str | Callable] | None,
-    margins: bool,
-    margins_name: str,
-    dropna: bool,
-    normalize: bool | str,
-) -> dict[str, DataFrame]:
-    """Create masks to specify the cells to suppress."""
-    # suppression masks to apply based on the following checks
-    masks: dict[str, DataFrame] = {}
-
-    if agg_func is not None:
-        # create lists with single entry for when there is only one aggfunc
-        count_funcs: list[str | Callable] = [AGGFUNC["count"]]
-        neg_funcs: list[Callable] = [agg_negative]
-        pperc_funcs: list[Callable] = [agg_p_percent]
-        nk_funcs: list[Callable] = [agg_nk]
-        missing_funcs: list[Callable] = [agg_missing]
-        # then expand them to deal with extra columns as needed
-        if isinstance(agg_func, list):
-            num = len(agg_func)
-            count_funcs.extend([AGGFUNC["count"] for i in range(1, num)])
-            neg_funcs.extend([agg_negative for i in range(1, num)])
-            pperc_funcs.extend([agg_p_percent for i in range(1, num)])
-            nk_funcs.extend([agg_nk for i in range(1, num)])
-            missing_funcs.extend([agg_missing for i in range(1, num)])
-        # threshold check- doesn't matter what we pass for value
-        if agg_func is mode_aggfunc:
-            # check that all observations dont have the same value
-            logger.info(
-                "If there are multiple modes, one of them is randomly selected and displayed."
-            )
-            masks["all-values-are-same"] = pd.crosstab(
-                index,
-                columns,
-                values,
-                aggfunc=agg_values_are_same,
-                margins=margins,
-                dropna=dropna,
-            )
-        else:
-            t_values = pd.crosstab(
-                index,
-                columns,
-                values=values,
-                rownames=rownames,
-                colnames=colnames,
-                aggfunc=count_funcs,
-                margins=margins,
-                margins_name=margins_name,
-                dropna=dropna,
-                normalize=normalize,
-            )
-
-            # drop empty columns and rows
-            if dropna or margins:
-                empty_cols_mask = t_values.sum(axis=0) == 0
-                empty_rows_mask = t_values.sum(axis=1) == 0
-
-                t_values = t_values.loc[:, ~empty_cols_mask]
-                t_values = t_values.loc[~empty_rows_mask, :]
-
-            t_values = t_values < THRESHOLD
-            masks["threshold"] = t_values
-            # check for negative values -- currently unsupported
-            negative = pd.crosstab(
-                index, columns, values, aggfunc=neg_funcs, margins=margins
-            )
-            if negative.to_numpy().sum() > 0:
-                masks["negative"] = negative
-            # p-percent check
-            masks["p-ratio"] = pd.crosstab(
-                index,
-                columns,
-                values,
-                aggfunc=pperc_funcs,
-                margins=margins,
-                dropna=dropna,
-            )
-            # nk values check
-            masks["nk-rule"] = pd.crosstab(
-                index, columns, values, aggfunc=nk_funcs, margins=margins, dropna=dropna
-            )
-            # check for missing values -- currently unsupported
-            if CHECK_MISSING_VALUES:
-                masks["missing"] = pd.crosstab(
-                    index, columns, values, aggfunc=missing_funcs, margins=margins
-                )
-    else:
-        # threshold check- doesn't matter what we pass for value
-        t_values = pd.crosstab(
-            index,
-            columns,
-            values=None,
-            rownames=rownames,
-            colnames=colnames,
-            aggfunc=None,
-            margins=margins,
-            margins_name=margins_name,
-            dropna=dropna,
-            normalize=normalize,
-        )
-        t_values = t_values < THRESHOLD
-        masks["threshold"] = t_values
-
-    # pd.crosstab returns nan for an empty cell
-    for name, mask in masks.items():
-        mask.fillna(value=1, inplace=True)
-        mask = mask.astype(int)
-        mask.replace({0: False, 1: True}, inplace=True)
-        masks[name] = mask
-    return masks
-
-
-def unused_delete_empty_rows_columns(table: DataFrame) -> tuple[DataFrame, list[str]]:
-    """Delete empty rows and columns from table.
-
-    Parameters
-    ----------
-    table : DataFrame
-        The table where the empty rows and columns will be deleted from.
-
-    Returns
-    -------
-    DataFrame
-        The resulting table where the empty columns and rows were deleted.
-    list[str]
-        A comment showing information about the deleted columns and rows.
-    """
-    deleted_rows: list[Any] = []
-    deleted_cols: list[Any] = []
-    # define empty columns and rows using boolean masks
-    empty_cols_mask: Any = table.sum(axis=0) == 0
-    empty_rows_mask: Any = table.sum(axis=1) == 0
-
-    deleted_cols = list(table.columns[empty_cols_mask])
-    table = table.loc[:, ~empty_cols_mask]
-    deleted_rows = list(table.index[empty_rows_mask])
-    table = table.loc[~empty_rows_mask, :]
-
-    # create a message with the deleted column's names
-    comments: list[str] = []
-    if deleted_cols:
-        msg_cols = ", ".join(str(col) for col in deleted_cols)
-        comments.append(f"Empty columns: {msg_cols} were deleted.")
-    if deleted_rows:
-        msg_rows = ", ".join(str(row) for row in deleted_rows)
-        comments.append(f"Empty rows: {msg_rows} were deleted.")
-    if comments:
-        logger.info(" ".join(comments))
-    return (table, comments)
-def unused_create_crosstab_masks(  # pylint: disable=too-many-arguments,too-many-locals
-    index: Any,
-    columns: Any,
-    values: Any,
-    rownames: Any,
-    colnames: Any,
-    agg_func: str | Callable | list[str | Callable] | None,
-    margins: bool,
-    margins_name: str,
-    dropna: bool,
-    normalize: bool | str,
-) -> dict[str, DataFrame]:
-    """Create masks to specify the cells to suppress."""
-    # suppression masks to apply based on the following checks
-    masks: dict[str, DataFrame] = {}
-
-    if agg_func is not None:
-        # create lists with single entry for when there is only one aggfunc
-        count_funcs: list[str | Callable] = [AGGFUNC["count"]]
-        neg_funcs: list[Callable] = [agg_negative]
-        pperc_funcs: list[Callable] = [agg_p_percent]
-        nk_funcs: list[Callable] = [agg_nk]
-        missing_funcs: list[Callable] = [agg_missing]
-        # then expand them to deal with extra columns as needed
-        if isinstance(agg_func, list):
-            num = len(agg_func)
-            count_funcs.extend([AGGFUNC["count"] for i in range(1, num)])
-            neg_funcs.extend([agg_negative for i in range(1, num)])
-            pperc_funcs.extend([agg_p_percent for i in range(1, num)])
-            nk_funcs.extend([agg_nk for i in range(1, num)])
-            missing_funcs.extend([agg_missing for i in range(1, num)])
-        # threshold check- doesn't matter what we pass for value
-        if agg_func is mode_aggfunc:
-            # check that all observations dont have the same value
-            logger.info(
-                "If there are multiple modes, one of them is randomly selected and displayed."
-            )
-            masks["all-values-are-same"] = pd.crosstab(
-                index,
-                columns,
-                values,
-                aggfunc=agg_values_are_same,
-                margins=margins,
-                dropna=dropna,
-            )
-        else:
-            t_values = pd.crosstab(
-                index,
-                columns,
-                values=values,
-                rownames=rownames,
-                colnames=colnames,
-                aggfunc=count_funcs,
-                margins=margins,
-                margins_name=margins_name,
-                dropna=dropna,
-                normalize=normalize,
-            )
-
-            # drop empty columns and rows
-            if dropna or margins:
-                empty_cols_mask = t_values.sum(axis=0) == 0
-                empty_rows_mask = t_values.sum(axis=1) == 0
-
-                t_values = t_values.loc[:, ~empty_cols_mask]
-                t_values = t_values.loc[~empty_rows_mask, :]
-
-            t_values = t_values < THRESHOLD
-            masks["threshold"] = t_values
-            # check for negative values -- currently unsupported
-            negative = pd.crosstab(
-                index, columns, values, aggfunc=neg_funcs, margins=margins
-            )
-            if negative.to_numpy().sum() > 0:
-                masks["negative"] = negative
-            # p-percent check
-            masks["p-ratio"] = pd.crosstab(
-                index,
-                columns,
-                values,
-                aggfunc=pperc_funcs,
-                margins=margins,
-                dropna=dropna,
-            )
-            # nk values check
-            masks["nk-rule"] = pd.crosstab(
-                index, columns, values, aggfunc=nk_funcs, margins=margins, dropna=dropna
-            )
-            # check for missing values -- currently unsupported
-            if CHECK_MISSING_VALUES:
-                masks["missing"] = pd.crosstab(
-                    index, columns, values, aggfunc=missing_funcs, margins=margins
-                )
-    else:
-        # threshold check- doesn't matter what we pass for value
-        t_values = pd.crosstab(
-            index,
-            columns,
-            values=None,
-            rownames=rownames,
-            colnames=colnames,
-            aggfunc=None,
-            margins=margins,
-            margins_name=margins_name,
-            dropna=dropna,
-            normalize=normalize,
-        )
-        t_values = t_values < THRESHOLD
-        masks["threshold"] = t_values
-
-    # pd.crosstab returns nan for an empty cell
-    for name, mask in masks.items():
-        mask.fillna(value=1, inplace=True)
-        mask = mask.astype(int)
-        mask.replace({0: False, 1: True}, inplace=True)
-        masks[name] = mask
-    return masks
-
-
-def unused_delete_empty_rows_columns(table: DataFrame) -> tuple[DataFrame, list[str]]:
-    """Delete empty rows and columns from table.
-
-    Parameters
-    ----------
-    table : DataFrame
-        The table where the empty rows and columns will be deleted from.
-
-    Returns
-    -------
-    DataFrame
-        The resulting table where the empty columns and rows were deleted.
-    list[str]
-        A comment showing information about the deleted columns and rows.
-    """
-    deleted_rows: list[Any] = []
-    deleted_cols: list[Any] = []
-    # define empty columns and rows using boolean masks
-    empty_cols_mask: Any = table.sum(axis=0) == 0
-    empty_rows_mask: Any = table.sum(axis=1) == 0
-
-    deleted_cols = list(table.columns[empty_cols_mask])
-    table = table.loc[:, ~empty_cols_mask]
-    deleted_rows = list(table.index[empty_rows_mask])
-    table = table.loc[~empty_rows_mask, :]
-
-    # create a message with the deleted column's names
-    comments: list[str] = []
-    if deleted_cols:
-        msg_cols = ", ".join(str(col) for col in deleted_cols)
-        comments.append(f"Empty columns: {msg_cols} were deleted.")
-    if deleted_rows:
-        msg_rows = ", ".join(str(row) for row in deleted_rows)
-        comments.append(f"Empty rows: {msg_rows} were deleted.")
-    if comments:
-        logger.info(" ".join(comments))
-    return (table, comments)
-
-def unused_get_aggfunc(aggfunc: str | None) -> str | Callable | None:
-    """Check whether an aggregation function is allowed and return the appropriate function.
-
-    Parameters
-    ----------
-    aggfunc : str | None
-        Name of the aggregation function to apply.
-
-    Returns
-    -------
-    Callable | None
-        The aggregation function to apply.
-    """
-    logger.debug("get_aggfunc()")
-    func: str | Callable | None = None
-    if aggfunc is not None:
-        if not isinstance(aggfunc, str):  # pragma: no cover
-            raise ValueError(f"aggfunc {aggfunc} must be:{', '.join(AGGFUNC.keys())}")
-        if aggfunc not in AGGFUNC:  # pragma: no cover
-            raise ValueError(f"aggfunc {aggfunc} must be: {', '.join(AGGFUNC.keys())}")
-        func = AGGFUNC[aggfunc]
-    logger.debug("aggfunc: %s", func)
-    return func
-
-def unused_get_aggfuncs(
-    aggfuncs: str | list[str] | None,
-) -> str | Callable | list[str | Callable] | None:
-    """Check whether aggregation functions are allowed and return appropriate functions.
-
-    Parameters
-    ----------
-    aggfuncs : str | list[str] | None
-        List of names of the aggregation functions to apply.
-
-    Returns
-    -------
-    Callable | list[Callable] | None
-        The aggregation functions to apply.
-    """
-    logger.debug("get_aggfuncs()")
-    if aggfuncs is None:
-        logger.debug("aggfuncs: None")
-        return None
-    if isinstance(aggfuncs, str):
-        function = get_aggfunc(aggfuncs)
-        logger.debug("aggfuncs: %s", function)
-        return function
-    if isinstance(aggfuncs, list):
-        functions: list[str | Callable] = []
-        for function_name in aggfuncs:
-            function = get_aggfunc(function_name)
-            if function is not None:
-                functions.append(function)
-        logger.debug("aggfuncs: %s", functions)
-        if len(functions) < 1:  # pragma: no cover
-            raise ValueError(f"invalid aggfuncs: {aggfuncs}")
-        return functions
-    raise ValueError("aggfuncs must be: either str or list[str]")  # pragma: no cover
-
-def unused__broadcast_mask_to_multiindex(
-    m: DataFrame, table: DataFrame, top_levels: list
-) -> DataFrame:
-    """Replicate a flat (or single-group) mask across each aggfunc top-level group.
-
-    Parameters
-    ----------
-    m : DataFrame
-        Flat mask with base columns only.
-    table : DataFrame
-        MultiIndex table whose top-level groups define the replication targets.
-    top_levels : list
-        Ordered list of unique top-level values from the table's MultiIndex columns.
-
-    Returns
-    -------
-    DataFrame
-        A mask with MultiIndex columns matching the table's column structure.
-    """
-    frames = []
-    for lvl in top_levels:
-        sub_cols = table[lvl].columns
-        sub_m = m.reindex(index=table.index, columns=sub_cols, fill_value=False)
-        sub_m.columns = pd.MultiIndex.from_product([[lvl], sub_m.columns])
-        frames.append(sub_m)
-    return pd.concat(frames, axis=1)
-
-
-def unused__align_mask_columns(m: DataFrame, table: DataFrame) -> DataFrame:
-    """Align the columns of mask *m* to match those of *table*.
-
-    Parameters
-    ----------
-    m : DataFrame
-        A single suppression mask.
-    table : DataFrame
-        The output table the mask should match.
-
-    Returns
-    -------
-    DataFrame
-        The mask with columns aligned to the table.
-    """
-    table_nlevels = table.columns.nlevels
-    mask_nlevels = m.columns.nlevels
-
-    if table_nlevels == 2 and mask_nlevels == 2:
-        table_top = table.columns.get_level_values(0).unique().tolist()
-        mask_top = m.columns.get_level_values(0).unique().tolist()
-        if mask_top != table_top:
-            n_base = len(table.columns.get_level_values(1).unique())
-            base_mask = m.iloc[:, :n_base]
-            flat_cols = base_mask.columns.get_level_values(1)
-            base_mask = pd.DataFrame(base_mask.values, index=m.index, columns=flat_cols)
-            m = _broadcast_mask_to_multiindex(base_mask, table, table_top)
-    elif mask_nlevels < table_nlevels:
-        top_levels = table.columns.get_level_values(0).unique().tolist()
-        m = _broadcast_mask_to_multiindex(m, table, top_levels)
-    elif mask_nlevels > table_nlevels:
-        m = m.droplevel(0, axis=1)
-
-    return m
-
-
-def unused_align_masks(table: DataFrame, masks: dict[str, DataFrame]) -> dict[str, DataFrame]:
-    """Align masks to the table's index and columns.
-
-    Parameters
-    ----------
-    table : DataFrame
-        Table to align masks to.
-    masks : dict[str, DataFrame]
-        Dictionary of tables specifying suppression masks.
-
-    Returns
-    -------
-    dict[str, DataFrame]
-        The aligned masks.
-    """
-    aligned_masks = {}
-    for name, mask in masks.items():
-        m = mask
-
-        # handle index level mismatch
-        if m.index.nlevels > table.index.nlevels:
-            m = m.droplevel(0, axis=0)
-
-        m = _align_mask_columns(m, table)
-
-        # reindex if still necessary
-        if not m.index.equals(table.index) or not m.columns.equals(table.columns):
-            try:
-                m = m.reindex(
-                    index=table.index, columns=table.columns, fill_value=False
-                )
-            except (ValueError, TypeError):  # pragma: no cover
-                logger.warning("Could not reindex mask %s", name)
-        aligned_masks[name] = m
-    return aligned_masks
-
-
-def unused_apply_suppression(
-    table: DataFrame, masks: dict[str, DataFrame]
-) -> tuple[DataFrame, DataFrame]:
-    """Apply suppression to a table.
-
-    Parameters
-    ----------
-    table : DataFrame
-        Table to apply suppression.
-    masks : dict[str, DataFrame]
-        Dictionary of tables specifying suppression masks for application.
-
-    Returns
-    -------
-    DataFrame
-        Table to output with any suppression applied.
-    DataFrame
-        Table with outcomes of suppression checks.
-    """
-    logger.debug("apply_suppression()")
-    # align masks
-    masks = align_masks(table, masks)
-    safe_df = table.copy()
-    outcome_df = DataFrame(index=table.index, columns=table.columns)
-    outcome_df.fillna("", inplace=True)
-    # don't apply suppression if negatives are present
-    if "negative" in masks:
-        mask = masks["negative"]
-        outcome_df[mask.values] = "negative"
-    # don't apply suppression if missing values are present
-    elif "missing" in masks:
-        mask = masks["missing"]
-        outcome_df[mask.values] = "missing"
-    # apply suppression masks
-    else:
-        for name, mask in masks.items():
-            try:
-                safe_df[mask.values] = np.nan
-                tmp_df = DataFrame(index=outcome_df.index, columns=outcome_df.columns)
-                tmp_df.fillna("", inplace=True)
-                tmp_df[mask.values] = name + "; "
-                outcome_df += tmp_df
-            except TypeError:
-                logger.warning("problem mask %s is not binary", name)
-            except ValueError as error:  # pragma: no cover
-                error_message = (
-                    f"An error occurred with the following details"
-                    f":\n Name: {name}\n Mask: {mask}\n Table: {table}"
-                )
-                raise ValueError(error_message) from error
-
-        outcome_df = outcome_df.replace({"": "ok"})
-    logger.info("outcome_df:\n%s", utils.prettify_table_string(outcome_df))
-    return safe_df, outcome_df
-
-def unused_get_queries_from_masks(
-    masks: dict[str, DataFrame],
-    aggfunc: str | Callable | list[str | Callable] | None,
-) -> list[str]:
-    """Return a list of the boolean conditions for each true (disclosive) cell in the suppression masks.
-
-    Parameters
-    ----------
-    masks : dict[str, DataFrame]
-        Dictionary of tables specifying suppression masks for application.
-    aggfunc : str | None
-        The aggregation function
-
-    Returns
-    -------
-    str
-        The boolean conditions for each true (disclosive) cell in the suppression masks.
-    """
-    true_cell_queries = []
-    for _, mask in masks.items():
-        if aggfunc is not None:
-            if mask.columns.nlevels > 1:
-                mask = mask.droplevel(0, axis=1)
-        index_level_names = mask.index.names
-        column_level_names = mask.columns.names
-        for col_index, _ in enumerate(mask.columns):
-            for row_index, _ in enumerate(mask.index):
-                query = _get_cell_query(
-                    mask, row_index, col_index, index_level_names, column_level_names
-                )
-                if query is not None:
-                    true_cell_queries.append(query)
-    true_cell_queries = list(set(true_cell_queries))
-    return true_cell_queries
-
-def unused_create_dataframe(index: Any, columns: Any) -> DataFrame:
-    """Combine the index and columns in a dataframe and return the dataframe.
-
-    Parameters
-    ----------
-    index : array-like, Series, or list of arrays/Series
-        Values to group by in the rows.
-    columns : array-like, Series, or list of arrays/Series
-        Values to group by in the columns.
-
-    Returns
-    -------
-    Dataframe
-        Table of the index and columns combined.
-    """
-    empty_dataframe = pd.DataFrame([])
-
-    index_df = empty_dataframe
-    try:
-        if isinstance(index, list):
-            index_df = pd.concat(index, axis=1)
-        elif isinstance(index, pd.Series):
-            index_df = pd.DataFrame({index.name: index})
-    except (ValueError, TypeError):
-        index_df = empty_dataframe
-
-    columns_df = empty_dataframe
-    try:
-        if isinstance(columns, list):
-            columns_df = pd.concat(columns, axis=1)
-        elif isinstance(columns, pd.Series):
-            columns_df = pd.DataFrame({columns.name: columns})
-    except (ValueError, TypeError):
-        columns_df = empty_dataframe
-
-    try:
-        data = pd.concat([index_df, columns_df], axis=1)
-    except (ValueError, TypeError):  # pragma: no cover
-        data = empty_dataframe
-
-    return data
