+++ acro/constants.py
@@ -1,0 +1,3 @@
+#https://www.sdmx.io/learning/essential-sdmx-structural-modelling/003-concept-mapping/
+SDMX_DIMENSION:str= "sdmx.model.Dimension"
+SDMX_MEASURE:str= "sdmx.model.Measure"
