+++ acro/utils.py
@@ -6,8 +6,15 @@ import logging
 from inspect import FrameInfo, getframeinfo
 
 import pandas as pd
-
+import numpy as np
 logger = logging.getLogger("acro")
+
+def get_catdtype(series:pd.Series)->pd.CategoricalDtype:
+    """ get info for pandas datatype to convert series to CategoricalDtype"""
+    ordered=  True if series.astype(int, errors="ignore").dtype=='int64' else False
+    categories= np.sort(series.unique())
+    cat_type= pd.CategoricalDtype(categories, ordered)
+    return cat_type
 
 
 def get_command(default: str, stack_list: list[FrameInfo]) -> str:
